shell: split windows and popups into different modules
