+++ src/shell/mod.rs
@@ -1,0 +1,4 @@
+pub mod popup;
+pub mod window;
+
+mod xdg_inner;
