+++ src/shell/popup/inner.rs
@@ -1,0 +1,6 @@
+use wayland_protocols::xdg_shell::client::xdg_popup;
+
+#[derive(Debug)]
+pub(crate) struct XdgPopupInner {
+    pub(crate) popup: xdg_popup::XdgPopup,
+}
