+++ src/shell/popup/mod.rs
@@ -1,0 +1,1 @@
+pub(crate) mod inner;
