+++ src/shell/xdg_inner.rs
@@ -1,0 +1,12 @@
+use std::sync::{Arc, Mutex};
+
+use super::{popup::inner::XdgPopupInner, window::inner::XdgToplevelInner};
+
+#[derive(Debug)]
+pub(crate) enum XdgSurfaceInner {
+    Window(Arc<Mutex<XdgToplevelInner>>),
+
+    Popup(Arc<Mutex<XdgPopupInner>>),
+
+    Uninit,
+}
